Move query formatting methods to util
