+++ citadel/indico_citadel/util.py
@@ -6,6 +6,7 @@
 # see the LICENSE file for more details.
 
 import asyncio
+import re
 from concurrent.futures.thread import ThreadPoolExecutor
 from functools import wraps
 
@@ -32,3 +33,51 @@ def parallelize(func, entries, batch_size=200):
             loop.run_until_complete(asyncio.gather(*tasks))
 
     return wrapper
+
+
+def format_query(query, placeholders):
+    """Format and split the query into keywords and placeholders.
+
+    https://cern-search.docs.cern.ch/usage/operations/#advanced-queries
+
+    :param query: search query
+    :param placeholders: placeholder whitelist
+    :returns escaped query
+    """
+    patt = r'({}):([^:"\s]+|".+")\s*'.format('|'.join(placeholders.keys()))
+    # Extract all placeholders
+    p = ['+{}:{}'.format(placeholders[x.group(1)], x.group(2))
+         for x in re.finditer(patt, query) if x.group(1) in placeholders]
+    # Escape keyword based arguments
+    query = escape(re.sub(patt, '', query))
+    return '{} {}'.format(' '.join(p), query).strip()
+
+
+def format_filters(params, filters, range_filters):
+    """Extract any special placeholder filter, such as ranges, from the query params.
+
+    https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-query-string-query.html#_ranges
+
+    :param params: The filter query params
+    :param filters: The filter whitelist
+    :param range_filters: The range filter whitelist
+    :returns: filters, extracted placeholders
+    """
+    _filters = {}
+    query = []
+    for k, v in params.items():
+        if k not in filters:
+            continue
+        if k in range_filters:
+            match = re.match(r'[[{].+ TO .+[]}]', v)
+            if match:
+                query.append('+{}:{}'.format(range_filters[k], v))
+            continue
+        _filters[k] = v
+    return _filters, ' '.join(query)
+
+
+def escape(query):
+    """Prepend all special ElasticSearch characters with a backslash."""
+    patt = r'([+\-=><!(){}[\]\^"~*?:\\\/]|&&|\|\|)'
+    return re.sub(patt, r'\\\1', query)
